comments

+++ engine/src/course/java/sdm/engine/Person.java
@@ -2,11 +2,11 @@ package course.java.sdm.engine;
 
 public class Person implements HasName {
 
-    private final int m_IDNumber; //TODO check that in system there is only 1 id with map
+    private final Long m_IDNumber; //TODO check that in system there is only 1 id with map
     private String m_Name;
 
 
-    public Person(int i_IDNumber,String i_Name) {
+    public Person(long i_IDNumber,String i_Name) {
         this.m_IDNumber = i_IDNumber;
         m_Name= i_Name;
     }
@@ -21,7 +21,7 @@ public class Person implements HasName {
         m_Name=Input;
     }
 
-    public int getIDNumber ()
+    public long getIDNumber ()
     {
         return m_IDNumber;
     }
